requirements.txt
# project requirements
requests
duckduckgo-search
python-dotenv
google-generativeai
